Add files via upload
The purpose of this project is to create a two-player game called “SOS” using the MINIMAX algorithm. Before the game begins, a square grid of at least 3x3 squares in size is drawn. Players take turns to add either an "S" or an "O" to any square, with no requirement to use the same letter each turn. The object of the game is for each player to attempt to create the straight sequence S-O-S among connected squares (either diagonally, horizontally, or vertically), and to create as many such sequences as they can. It is a similar game to tic-tac-toe and dots and boxes, but has greater complexity.